Changed datapoint value from double to float (trying to fix json encoding issue...)

+++ TestWebApi/DataLogController.cs
@@ -13,7 +13,7 @@ namespace TestWebApi
         public class datapoint
         {
             public int time;
-            public double value;
+            public float value;
         }
 
         public static List<datapoint> datapoints = InitDatapoints();
